refactor(p1): extract WebMount Group C — Fetch + Site mgmt; coordinator slims to 110 lines
Closes WebMount god-class split: WebMountPrimitiveTools.kt 1604 → 110 lines (-93%).

Two new sibling files:
- WebMountFetchTools.kt — createSignedFetchTool(deps, profileRegistry, profileBridge).
  Single tool; the only WebMount primitive that goes through ProfileBridge.callSign.
  Preserves the B-2 fix (requestedUrlHost threaded through to enforce send_signed:<host>).
- WebMountSiteTools.kt — createStationsTool / createSiteAddTool / createSiteRemoveTool
  / createProfileSynthesizeTool. stations + remove take the full 6-dep set
  (userSiteRegistry + manager + profileRegistry + cookieProvider + oauthStore + deps);
  site_add only needs userSiteRegistry; profile_synthesize needs userSiteRegistry +
  profileRegistry. Three-state login signal (logged_in / logged_out / unknown)
  preserved verbatim, including the W-1 "drop enabled flag" / B-3 "profile fallback
  via site.id" / B-4 "OAuth needs_login parity" / B-5 "unconditional cookie+token
  clear" fixes from earlier reviews.

Coordinator rewritten as thin shell:
- 8-dep ctor stays (every dep is still consumed somewhere — pool/activityStore via
  WebMountDeps; manager/profileRegistry/cookieProvider via navigation + sites;
  profileBridge via fetch; userSiteRegistry/oauthStore via sites)
- single WebMountDeps wrapper field
- 23 `by lazy { createXxxTool(...) }` delegators
- getTools(includeEval) registry with original tool order preserved

Imports: dropped 13 imports that travelled with their tools (JsonNull, JsonObject,
buildJsonArray, buildJsonObject, kotlinx.serialization.json.put, contentOrNull,
jsonObject, InputSchema, UIMessagePart, several agent/tools/ + webmount/profile/
+ webmount/usersites/ symbols, SessionHandle, WebViewScreenshot, etc.).

Cumulative result (Sessions #3+#4 P1 god-class work):
- LocalTools.kt: 1093 → 189 lines  (Session #3, 14 sibling files, -83%)
- SystemAccessTools.kt: 1441 → 93 lines  (Session #4, 13 sibling files, -93%)
- WebMountPrimitiveTools.kt: 1604 → 110 lines  (Session #4, 7 sibling files, -93%)

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
